Player Weapon and Item Table update!
Added functional tables which are editable by the user in the "Players"
tab on the mainGUI. User can add rows as well as remove selected rows at
their desire. To edit a cell just double click. This will act as a way
for the user to edit on the fly.